pyproject.toml
[tool.poetry]
name = "HM_db_connector"
version = "0.1.0"
description = "Protocol-agnostic request library that unifies HTTP, CoAP, MQTT and future transports under a single API."
authors = ["sato96 <samtozzi.st@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "HM_db_connector", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
